Alterations to how canvas displays the image
A few big changes here!
I have mostly kept the code the same (and broken parts along the way)…
- Rather than using CSS filters we need to apply them directly to the
canvas, so that we can retrieve a dataURL which holds those filters in
it.
- Luckily they just released changes to the canvas (2 days ago) which
allow CSS style filters, so I have reworked the code into seperate
functions to get these filters working
- You need to use the canvas image display instead of the static image
on the right hand side
- Filter adjustment is not working in totally the same way, just needs
some tweaks
- You need to wait for the image dataURL to be saved to localStorage
before going off to the next page, so you might need to put some kind
of loading screen in while you wait for it to save
- The save image script will only work on a server, I’ve put it on my
dev server (link in the top of the php file) note that it saves
something, but it is empty! I will play a bit further here!
- Lots of comments added, happy to explain in person if all a bit
confusing! I will try get the saveImage part to actually work a bit
later.